requirements.txt
kafka-python==2.0.2
pandas==2.2.2
pyspark==3.5.1
prometheus-client==0.20.0
